requirements.txt
requests==2.20.0
httmock==1.3.0
newspaper3k==0.2.8
tests==0.007